fix c3d bug

+++ mmaction/models/data_preprocessors/data_preprocessor.py
@@ -55,10 +55,10 @@ class ActionDataPreprocessor(BaseDataPreprocessor):
                 raise ValueError(f'Invalid format shape: {format_shape}')
 
             self.register_buffer('mean',
-                                 torch.tensor(mean).view(normalizer_shape),
+                                 torch.tensor(mean, dtype=torch.float32).view(normalizer_shape),
                                  False)
             self.register_buffer('std',
-                                 torch.tensor(std).view(normalizer_shape),
+                                 torch.tensor(std, dtype=torch.float32).view(normalizer_shape),
                                  False)
         else:
             self._enable_normalize = False
